safari-600.7-branch/LayoutTests/fast/viewport/viewport-112-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "." not recognized and ignored.
ALERT: viewport size 400x440 scale 0.8 with limits [0.8, 5] and userScalable true
